Split auth on frontend into AuthContext, AuthProvider, and useAuth
