+++ frontend/src/contexts/AuthContext.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const AuthContext = React.createContext();
