+++ frontend/src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react';
 import { BrowserRouter } from "react-router-dom";
 import { createRoot } from 'react-dom/client';
-import App from './App.jsx';
-import { AuthProvider } from './contexts/AuthContext.jsx';
+import App from './App';
+import { AuthProvider } from './contexts/AuthProvider';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
